backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.8
dj-stripe==2.5.1
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.10
jsonfield==3.1.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
urllib3==1.26.6
